FUN-24 feat: remove pages
