+++ src/pages/home/index.tsx
@@ -1,6 +1,5 @@
 import { getProtectedNavbarLayout } from '@/shared/layouts'
 import { HeadMeta } from '@/shared/ui'
-import { EmailVerification } from '@/widgets/EmailVerification/ui/EmailVerification'
 
 const HomePage = () => (
   <>
