+++ src/pages/index.tsx
@@ -1,49 +1,26 @@
-import { GetPostResponse, GetPostsResponse } from '@/shared/api/postsApi'
-import { BASE_API_URL } from '@/shared/const'
+import { useEffect } from 'react'
+
+import { BASE_LOCAL_URL, BASE_URL } from '@/shared/const'
 import { getRootLayout } from '@/shared/layouts'
-import { HeadMeta, PublicPosts, Typography } from '@/shared/ui'
-import { TotalUsersCounter } from '@/widgets/TotalUsersCounter'
-import { GetStaticProps } from 'next'
+import { GoogleAuth } from '@/widgets/GoogleAuth'
+import { useRouter } from 'next/router'
 
-export const getStaticProps: GetStaticProps = async () => {
-  const res = await fetch(`${BASE_API_URL}public-posts/all?pageSize=4`)
-  const posts: GetPostsResponse = await res.json()
+const Page = () => {
+  const router = useRouter()
+  const { code } = router.query as { code: string }
 
-  return {
-    props: {
-      error: 'No posts yet!',
-      publicPosts: posts.items,
-      totalUsersCount: posts.totalUsers,
-    },
-    revalidate: 60,
+  if (code) {
+    return <GoogleAuth code={code} />
   }
+
+  useEffect(() => {
+    if (!code) {
+      router.push('/profile')
+    }
+  }, [router])
+
+  return null
 }
 
-type PropsType = {
-  error: string
-  publicPosts: GetPostResponse[]
-  totalUsersCount: number
-}
-
-const PublicPage = ({ error, publicPosts, totalUsersCount }: PropsType) => {
-  return (
-    <>
-      <HeadMeta title={'public page'} />
-      {publicPosts ? (
-        <main style={{ margin: '0 auto', maxWidth: '972px' }}>
-          <TotalUsersCounter usersCount={totalUsersCount} />
-          {publicPosts && <PublicPosts publicPosts={publicPosts} />}
-        </main>
-      ) : (
-        <div style={{ paddingTop: '36px', textAlign: 'center' }}>
-          <Typography as={'p'} variant={'boldText16'}>
-            {error}
-          </Typography>
-        </div>
-      )}
-    </>
-  )
-}
-
-PublicPage.getLayout = getRootLayout
-export default PublicPage
+Page.getLayout = getRootLayout
+export default Page
